+++ housekeeping-vacuum-tool/src/main/java/com/hotels/housekeeping/tool/vacuum/VacuumTool.java
@@ -40,19 +40,21 @@ public class VacuumTool {
     logVersionInfo();
 
     try {
-      SpringApplication.exit(new SpringApplicationBuilder(VacuumTool.class)
-          .properties("spring.config.location:${config:null}")
-          .properties("instance.name:${catalog.name}")
-          .properties("instance.home:${user.home}")
-          .registerShutdownHook(true)
-          .bannerMode(Mode.OFF)
-          .registerShutdownHook(true)
-          .listeners(new ConfigFileValidationApplicationListener())
-          .build()
-          .run(args));
+      SpringApplication
+          .exit(new SpringApplicationBuilder(VacuumTool.class)
+              .properties("spring.config.location:${config:null}")
+              .properties("instance.name:${catalog.name}")
+              .properties("instance.home:${user.home}")
+              .registerShutdownHook(true)
+              .bannerMode(Mode.OFF)
+              .registerShutdownHook(true)
+              .listeners(new ConfigFileValidationApplicationListener())
+              .build()
+              .run(args));
     } catch (BeanCreationException e) {
-      if (e.getMostSpecificCause() instanceof BindException) {
-        printVacuumToolHelp(((BindException) e.getMostSpecificCause()).getAllErrors());
+      Throwable mostSpecificCause = e.getMostSpecificCause();
+      if (mostSpecificCause instanceof BindException) {
+        printVacuumToolHelp(((BindException) mostSpecificCause).getAllErrors());
       }
       throw e;
     }