+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [TBD]
+### Fixed
+* Fixed deletion of path entries in the database when `pathEventId` is null. See [#48](https://github.com/HotelsDotCom/housekeeping/issues/48).
+
+## [3.0.1] - 2019-01-10
 ### Changed
 * Refactored housekeeping and vacuum-tool to remove checkstyle and findbugs warnings, which does not impact functionality.
 * Upgraded `hotels-oss-parent` to 2.3.5 (was 2.3.3).